+++ src/components/Theme.tsx
@@ -1,8 +1,8 @@
-import { createContext, useState, useContext } from "react";
+import { createContext, useState} from "react";
 import type { ReactNode } from "react"
 import type { ThemeContextType } from "@/types/todo.types";
 
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
 
 interface ThemeProviderProps {
   children: ReactNode;
@@ -20,12 +20,4 @@ export function ThemeProvider({ children }: ThemeProviderProps) {
       {children}
     </ThemeContext.Provider>
   );
-}
-
-export function useTheme(): ThemeContextType {
-  const context = useContext(ThemeContext);
-  if (!context) {
-    throw new Error("useTheme must be used within a ThemeProvider");
-  }
-  return context;
 }
